Consolidate regx launch helpers
